+++ webpack.config.prod.babel.js
@@ -1,7 +1,7 @@
 import path from 'path';
 import webpack from 'webpack';
 import ExtractTextPlugin from 'extract-text-webpack-plugin';
-import MappingPlugin from 'webpack-mapping-plugin';
+import ManifestPlugin from 'webpack-manifest-plugin';
 import autoprefixer from 'autoprefixer';
 import flexbugs from 'postcss-flexbugs-fixes'; // 修复 flexbox 已知的 bug
 //const cssnano = require('cssnano'); // 优化 css，对于长格式优化成短格式等
@@ -59,15 +59,16 @@ function scssConfig(modules) {
     use: [{
       loader: 'css-loader',
       options: modules ? {
-        // sourceMap: true,
+        sourceMap: true,
         modules: true,
         minimize: true,
-        localIdentName: '[local][hash:base64:5]',
+        localIdentName: '[local][contenthash:base64:5]',
         getLocalIdent: (context, localIdentName, localName) => {
+          // FIXME 这样每次打包编译时，混淆的 css 名导致无法缓存，待解决实现
           return generateScopedName(localName, context.resourcePath);
         },
       } : {
-        // sourceMap: true,
+        sourceMap: true,
         minimize: true,
       },
     }, {
@@ -99,15 +100,16 @@ function scssConfig(modules) {
 
 // multiple extract instances
 const extractScss = new ExtractTextPlugin({
-  filename: 'css/[name].[chunkhash].css',
+  filename: 'css/[name].[contenthash].css',
   allChunks: true,
   ignoreOrder: true,
 });
 const extractCSS = new ExtractTextPlugin({
-  filename: 'css/style.[name].[chunkhash].css',
+  filename: 'css/style.[name].[contenthash].css',
   allChunks: true,
 });
 
+// 基于 webpack 的持久化缓存方案 可以参考 https://github.com/pigcan/blog/issues/9
 const webpackConfig = {
   devtool: 'source-map', // 生成 source-map文件 原始源码
   target: 'web', // webpack 能够为多种环境构建编译, 默认是 'web'，可省略 https://doc.webpack-china.org/configuration/target/
@@ -211,24 +213,32 @@ const webpackConfig = {
   },
 
   plugins: [
-    //用来优化生成的代码 chunk,合并相同的代码
+    // 用来优化生成的代码 chunk，合并相同的代码
     new webpack.optimize.AggressiveMergingPlugin(),
-    //用来保证编译过程不出错
+    // 用来保证编译过程不出错
     new webpack.NoEmitOnErrorsPlugin(),
-    new webpack.optimize.ModuleConcatenationPlugin(), // Scope Hoisting-作用域提升
+    new webpack.optimize.ModuleConcatenationPlugin({// Scope Hoisting-作用域提升
+      // 检查所有的模块
+      maxModules: Infinity,
+      // 将显示绑定失败的原因
+      optimizationBailout: true,
+    }),
     new webpack.DefinePlugin({
       'process.env': {
         NODE_ENV: JSON.stringify('production'),
       },
     }),
-    // https://doc.webpack-china.org/guides/code-splitting-libraries/#manifest-
-    new webpack.optimize.CommonsChunkPlugin({
-      names: ['vendor', 'manifest'],
-      filename: '[name].[chunkhash].js',
-    }),
     extractScss,
     extractCSS,
-    new MappingPlugin({
+    // 以下两个插件可以解决持久化缓存，但由于用到了模块化，混淆的 css 名导致无法缓存
+    // 故先注释掉，待后续解决了，再打开，或者不使用 css Module 时
+    // new webpack.HashedModuleIdsPlugin(),
+    // new webpack.NamedChunksPlugin(),
+    // https://doc.webpack-china.org/guides/code-splitting-libraries/#manifest-
+    new webpack.optimize.CommonsChunkPlugin({
+      name: ['vendor', 'manifest'], // 或 runtime
+    }),
+    new ManifestPlugin({
       basePath: `${urlContext}/dist/`,
     }),
     new webpack.optimize.UglifyJsPlugin({
@@ -242,7 +252,16 @@ const webpackConfig = {
         except: [] // 设置不混淆变量名
       },
     }),
-    // new webpack.BannerPlugin({banner: 'Banner', raw: true, entryOnly: true}),
+    new webpack.BannerPlugin({
+      banner: [
+        '/*!',
+        ' react-redux-router-base',
+        ` Copyright © 2017-${new Date().getFullYear()} JD Finance.`,
+        '*/',
+      ].join('\n'),
+      raw: true,
+      entryOnly: true,
+    }),
     new webpack.LoaderOptionsPlugin({
       /*UglifyJsPlugin 不再压缩 loaders。在未来很长一段时间里，需要通过设置 minimize:true 来压缩 loaders。
        loaders 的压缩模式将在 webpack 3 或后续版本中取消。